rename fetching and database files
