+++ command_functions/make_private_command.js
@@ -9,7 +9,7 @@ const {
   isDiscordUserAdmin,
 } = require("../data/mongoDb");
 
-const { handleAPIError } = require("../fetching/errorHandling");
+const { handleAPIError } = require("../fetching/error_handling");
 
 const {
   generateLoadingTime,
