+++ command_functions/rank_command.js
@@ -9,7 +9,7 @@ const {
   removeLoadingInstance,
 } = require("../fetching/loading");
 
-const { handleAPIError } = require("../fetching/errorHandling");
+const { handleAPIError } = require("../fetching/error_handling");
 
 const {
   mmrDataSingleToString,
