+++ data/schemas/discord_user_schema.js
